sdcard: copy zImage/dtb to disk

+++ scripts/format_drive.sh
@@ -8,8 +8,13 @@ if [ "x${MMC}" = "x" ] ; then
 fi
 
 sudo dd if=/dev/zero of=${MMC} bs=1M count=10
-sudo dd if=./deploy/MLO of=${MMC} count=1 seek=1 bs=128k
-sudo dd if=./deploy/u-boot.img of=${MMC} count=2 seek=1 bs=384k
+
+if [ -f ./deploy/MLO ] ; then
+	sudo dd if=./deploy/MLO of=${MMC} count=1 seek=1 bs=128k
+fi
+if [ -f ./deploy/u-boot.img ] ; then
+	sudo dd if=./deploy/u-boot.img of=${MMC} count=2 seek=1 bs=384k
+fi
 
 sudo sfdisk ${MMC} <<-__EOF__
 4M,,L,*
@@ -25,7 +30,26 @@ sudo mount ${MMC}1 /media/rootfs/
 
 sleep 5
 
-sudo tar xfvp ./rootfs/debian-*/armhf-rootfs-*.tar -C /media/rootfs/
+if [ -f ./rootfs/debian-*/armhf-rootfs-*.tar ] ; then
+	sudo tar xfvp ./rootfs/debian-*/armhf-rootfs-*.tar -C /media/rootfs/
+	sync
+	sudo chown root:root /media/rootfs/
+	sudo chmod 755 /media/rootfs/
+
+	sudo sh -c "echo 'uname_r=4.14.0-rc7-bone3' > /media/rootfs/boot/uEnv.txt"
+	sudo sh -c "echo '/dev/mmcblk0p1  /  auto  errors=remount-ro  0  1' > /media/rootfs/etc/fstab"
+
+	if [ -f ./deploy/zImage ] ; then
+		sudo cp -v ./deploy/zImage /media/rootfs/boot/vmlinuz-4.14.0-rc7-bone3
+	fi
+
+	if [ -f ./deploy/am335x-pocketbeagle.dtb ] ; then
+		if [ ! -d /media/rootfs/boot/dtbs/4.14.0-rc7-bone3/ ] ; then
+			sudo mkdir -p /media/rootfs/boot/dtbs/4.14.0-rc7-bone3/
+		fi
+		sudo cp -v ./deploy/am335x-pocketbeagle.dtb /media/rootfs/boot/dtbs/4.14.0-rc7-bone3/
+	fi
+fi
+
 sync
-sudo chown root:root /media/rootfs/
-sudo chmod 755 /media/rootfs/
+sudo umount /media/rootfs
